Updated pcfg_productivity_data.py
Updated pcfg_productivity_data.py to generate both the original and the iterative decoding data. After running $ tfds build pcfg_productivity_data.py to build the dataset, the original data can be found in the 'train' and 'test' splits; and the iterative decoding data, in the 'it_dec_train', 'it_dec_val' and 'it_dec_test' splits. Per Josh's suggestion, the 'it_dec_val' set is used to assess the ability of the model to generalize in the standard sense (it has the same form as the 'it_dec_train' set); the 'it_dec_test' set is used to assess the ability of the model to generalize through iterative decoding.