Updated data_generation.py to output number of ops
data_generation.py now produces an .ops file for the it_dec_test split containing the number of operations in each input sequence of the iterative decoding test split. The number of operations is used to inform the number of prediction loops necessary to decode the iterative decoding test samples.